# bismarck-game/backend/docs/swagger.yaml
basePath: /api
definitions:
...
  internal_api_handlers.CheckEmergencyFuelRequest:
    properties:
      game_id:
        type: string
      unit_id:
        type: string
    type: object
  internal_api_handlers.CheckEmergencyFuelResponse:
    properties:
      current_fuel:
        type: integer
      emergency_turn:
        type: integer
      is_emergency_fuel:
        type: boolean
      message:
        type: string
      success:
        type: boolean
    type: object
  internal_api_handlers.RefuelAllRequest:
    properties:
      fuel_amount:
        type: integer
      game_id:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    name: Bismarck Game Team
  description: API для игры Bismarck Chase - пошаговой стратегической игры о морских
    сражениях
  title: Bismarck Game API
  version: "1.0"
paths:
...
  /emergency-fuel/check:
    post:
      consumes:
      - application/json
      parameters:
      - description: Данные для проверки аварийного топлива
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_api_handlers.CheckEmergencyFuelRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api_handlers.CheckEmergencyFuelResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Проверка и активация аварийного топлива
      tags:
      - Emergency Fuel
  /emergency-fuel/status:
    get:
      consumes:
      - application/json
      parameters:
      - description: ID игры
        in: query
        name: game_id
        required: true
        type: string
      - description: ID юнита
        in: query
        name: unit_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api_handlers.CheckEmergencyFuelResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Получение статуса аварийного топлива
      tags:
      - Emergency Fuel
...
  /refuel/all:
    post:
      consumes:
      - application/json
      parameters:
      - description: Данные для заправки
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_api_handlers.RefuelAllRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Заправка всех кораблей в игре
      tags:
      - Refuel
...
securityDefinitions:
  Bearer:
    description: 'JWT токен в формате: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"